Adicionando alert

+++ src/app/pages/list-page/list-page.component.ts
@@ -1,9 +1,11 @@
-import { Component, inject, OnInit, signal } from '@angular/core';
+import { Component, OnInit, signal } from '@angular/core';
 import { InputAddItemComponent } from '../../components/input-add-item/input-add-item.component';
 import { ITask } from '../../intefaces/task.interface';
 import { ResponseAPI, TaskService } from '../../service/task.service';
 import { NgClass } from '@angular/common';
 import { InputListItemComponent } from '../../components/input-list-item/input-list-item.component';
+import Swal, { SweetAlertResult } from 'sweetalert2';
+import { AlertService } from '../../service/alert.service';
 
 @Component({
   selector: 'app-list-page',
@@ -15,7 +17,10 @@ import { InputListItemComponent } from '../../components/input-list-item/input-l
 export class ListPageComponent implements OnInit {
   addItem = signal<boolean>(true);
   taskList: ITask[] = [];
-  constructor(private readonly taskService: TaskService) {}
+  constructor(
+    private readonly taskService: TaskService,
+    private alertService: AlertService
+  ) {}
 
   public getTasks() {
     this.taskService.getTasks().subscribe((data: ITask[]) => {
@@ -46,21 +51,36 @@ export class ListPageComponent implements OnInit {
   }
 
   public deleteTask(taskId: number) {
-    this.taskService.deleteTask(taskId).subscribe({
-      next: () => this.getTasks(),
-      error: (error) => {
-        console.log(error);
-      },
-    });
+    this.alertService
+      .openModal(`Apagar Tarefa`, `Voce realmente deseja apagar a tarefa ?`)
+      .then((result: SweetAlertResult) => {
+        if (result.isConfirmed) {
+          this.taskService.deleteTask(taskId).subscribe({
+            next: () => this.getTasks(),
+            error: (error) => {
+              console.log(error);
+            },
+          });
+        }
+      });
   }
 
   public deleteAllTasks() {
-    this.taskService.deleteAllTasks().subscribe({
-      next: () => this.getTasks(),
-      error: (error) => {
-        console.log(error);
-      },
-    });
+    this.alertService
+      .openModal(
+        `Apagar Todas as Tarefas`,
+        `Esta ação ira remover todas as suas tarefas, deseja continuar?`
+      )
+      .then((result: SweetAlertResult) => {
+        if (result.isConfirmed) {
+          this.taskService.deleteAllTasks().subscribe({
+            next: () => this.getTasks(),
+            error: (error) => {
+              console.log(error);
+            },
+          });
+        }
+      });
   }
 
   ngOnInit(): void {
